Extract DashboardPage component
